Prediction components renamed to Stake/Staking
