+++ src/components/Proposal/Staking/BoostAmount.scss
