+++ src/components/Proposal/Staking/BoostAmount.tsx
