+++ src/components/Proposal/Staking/StakeButtons.scss
