+++ src/components/Proposal/Staking/StakeGraph.scss
